+++ nodewallet/eth/eth.go
@@ -133,7 +133,7 @@ func (w *Wallet) Algo() string {
 	return "eth"
 }
 
-func (w *Wallet) Version() uint64 {
+func (w *Wallet) Version() uint32 {
 	return 0
 }
